update: small refactor
